// Nurture-Glow/frontend/types.ts
export interface JournalEntry {
  id: string;
  userId: string;
  title: string;
  content: string;
  date: string;
  mood?: string;
  attachments: {
    name: string;
    url: string; // Base64
    type: string;
  }[];
}